fsbackend/requirements.txt
djangorestframework
markdown
django-filter
django-cors-headers
drf-spectacular
drf_spectacular_sidecar
psycopg2-binary
requests
python-dotenv